+++ contrib/cachetool.php
@@ -96,3 +96,11 @@ desc('Clearing APCu system cache');
 task('cachetool:clear:apcu', function () {
     run("cd {{release_path}} && {{bin/php}} {{bin/cachetool}} apcu:cache:clear {{cachetool_options}}");
 });
+
+/**
+ * Clear file status cache, including the realpath cache
+ */
+desc('Clearing file status and realpath caches');
+task('cachetool:clear:stat', function () {
+    run("cd {{release_path}} && {{bin/php}} {{bin/cachetool}} stat:clear {{cachetool_options}}");
+});